fix(findings): stop exporting helpers from occurrences list and sync api-types (#7304)
- Extract the shared finding-occurrence helpers (occurrenceTargets,
  occurrenceTargetLink, buildOccurrencesFilter) into FindingOccurrencesUtils so
  FindingOccurrencesList only exports its component; this clears the
  react-refresh/only-export-components lint failure. Repoint the timeline import.
- Regenerate api-types.d.ts so the custom-domain and finding-summary schemas
  match the backend, fixing the API Types Check.
